app/core/law_translate/requirements.txt
googletrans==4.0.0rc1
python-docx
watchdog
requests
pydantic
pymupdf
fitz
sqlalchemy